# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.9.0"
  },
  "paths": {
...
    "/sampling/approve": {
      "post": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "handle_sampling_approval",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sampling approval processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sampling/request": {
      "post": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "handle_sampling_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sampling request received",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Content": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawTextContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawImageContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawEmbeddedResource"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawAudioContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawResource"
              }
            ]
          }
        ]
      },
...
      "Icon": {
        "type": "object",
        "required": [
          "src"
        ],
        "properties": {
          "mimeType": {
            "type": "string"
          },
          "sizes": {
            "type": "string"
          },
          "src": {
            "type": "string"
          }
        }
      },
...
      "RawAudioContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
      "RawEmbeddedResource": {
        "type": "object",
        "required": [
          "resource"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceContents"
          }
        }
      },
      "RawImageContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
      "RawResource": {
        "type": "object",
        "required": [
          "uri",
          "name"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "icons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Icon"
            }
          },
          "mimeType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          }
        }
      },
      "RawTextContent": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "ResourceContents": {
        "anyOf": [
          {
            "type": "object",
            "required": [
              "uri",
              "text"
            ],
            "properties": {
              "_meta": {
                "type": "object",
                "additionalProperties": true
              },
              "mimeType": {
                "type": "string"
              },
              "text": {
                "type": "string"
              },
              "uri": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "uri",
              "blob"
            ],
            "properties": {
              "_meta": {
                "type": "object",
                "additionalProperties": true
              },
              "blob": {
                "type": "string"
              },
              "mimeType": {
                "type": "string"
              },
              "uri": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "SamplingApprovalRequest": {
        "type": "object",
        "required": [
          "session_id",
          "original_request",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "The user's action: approve, deny, or edit"
          },
          "edited_messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SamplingMessage"
            },
            "description": "If action is \"edit\", this contains the edited messages",
            "nullable": true
          },
          "original_request": {
            "$ref": "#/components/schemas/SamplingRequest"
          },
          "session_id": {
            "type": "string",
            "description": "The session ID for the current agent session"
          }
        }
      },
      "SamplingMessage": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "$ref": "#/components/schemas/Content"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "SamplingRequest": {
        "type": "object",
        "required": [
          "session_id",
          "extension_name",
          "messages",
          "max_tokens"
        ],
        "properties": {
          "extension_name": {
            "type": "string",
            "description": "The extension name that is making the sampling request"
          },
          "include_context": {
            "type": "string",
            "description": "Whether to include context",
            "nullable": true
          },
          "max_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum tokens to generate"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SamplingMessage"
            },
            "description": "The messages to send to the LLM"
          },
          "metadata": {
            "description": "Additional metadata",
            "nullable": true
          },
          "model_preferences": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional model preferences",
            "nullable": true
          },
          "session_id": {
            "type": "string",
            "description": "The session ID for the current agent session"
          },
          "stop_sequences": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Stop sequences",
            "nullable": true
          },
          "system_prompt": {
            "type": "string",
            "description": "Optional system prompt",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "Temperature for the model",
            "nullable": true
          }
        }
      },
      "SamplingResponse": {
        "type": "object",
        "required": [
          "message",
          "model"
        ],
        "properties": {
          "message": {
            "$ref": "#/components/schemas/SamplingMessage"
          },
          "model": {
            "type": "string",
            "description": "The model used for generation"
          },
          "stop_reason": {
            "type": "string",
            "description": "The reason for stopping",
            "nullable": true
          }
        }
      },
...
    }
  }
}